[netdrv] i40e: Change unmatched function types
Message-id: <20190222143647.8024-80-sassmann@redhat.com>
Patchwork-id: 242927
O-Subject: [RHEL7.7 PATCH 79/84] i40e: Change unmatched function types
Bugzilla: 1637726
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

From: Grzegorz Siwik <grzegorz.siwik@intel.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1637726
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=20326389

Change of function declaration from int to u64 due to
return type mismatch (u64).

Tested-by: Andrew Bowers <andrewx.bowers@intel.com>

(cherry picked from commit 9b0732d9ed6f80d712119fa32ff25ae9b09d481e)

+++ drivers/net/ethernet/intel/i40e/i40e_ethtool.c
@@ -2372,7 +2372,7 @@ static int i40e_get_ts_info(struct net_device *dev,
 	return 0;
 }
 
-static int i40e_link_test(struct net_device *netdev, u64 *data)
+static u64 i40e_link_test(struct net_device *netdev, u64 *data)
 {
 	struct i40e_netdev_priv *np = netdev_priv(netdev);
 	struct i40e_pf *pf = np->vsi->back;
@@ -2395,7 +2395,7 @@ static int i40e_link_test(struct net_device *netdev, u64 *data)
 	return *data;
 }
 
-static int i40e_reg_test(struct net_device *netdev, u64 *data)
+static u64 i40e_reg_test(struct net_device *netdev, u64 *data)
 {
 	struct i40e_netdev_priv *np = netdev_priv(netdev);
 	struct i40e_pf *pf = np->vsi->back;
@@ -2406,7 +2406,7 @@ static int i40e_reg_test(struct net_device *netdev, u64 *data)
 	return *data;
 }
 
-static int i40e_eeprom_test(struct net_device *netdev, u64 *data)
+static u64 i40e_eeprom_test(struct net_device *netdev, u64 *data)
 {
 	struct i40e_netdev_priv *np = netdev_priv(netdev);
 	struct i40e_pf *pf = np->vsi->back;
@@ -2420,7 +2420,7 @@ static int i40e_eeprom_test(struct net_device *netdev, u64 *data)
 	return *data;
 }
 
-static int i40e_intr_test(struct net_device *netdev, u64 *data)
+static u64 i40e_intr_test(struct net_device *netdev, u64 *data)
 {
 	struct i40e_netdev_priv *np = netdev_priv(netdev);
 	struct i40e_pf *pf = np->vsi->back;
